Pushing Bonfire and FPS Loggger
I have made a FPSLogger script (attached to a logging manager), which should save a file on the headset with the logging data once the experience is closed (still needs testing). I also adjusted the bonfire, so that the logs and fire works with the BuildingBlock Camera Rig. Might need some fine tweaking, but overall looks good (i think)